+++ Tools/SkinEditor/StateTextureControl.cpp
@@ -16,7 +16,7 @@ namespace tools
 		mAreaSelectorControl(nullptr),
 		mGridStep(0)
 	{
-		mTypeName = MyGUI::utility::toString((int)this);
+		mTypeName = MyGUI::utility::toString((size_t)this);
 
 		// сразу рисуем рамки для стейтов
 		std::vector<MyGUI::IntCoord> coords(8);
